# Bug: distance_to_space on wrong thing. Massive performance wins.