This is big! I stumbled on this experimenting with a refactoring. 

New bench:
```
running 2 tests
test bench_add_to_kdtree_with_1k_3d_points       ... bench:         156 ns/iter (+/- 44)
test bench_nearest_from_kdtree_with_1k_3d_points ... bench:       3,603 ns/iter (+/- 3,217)
```

Old bench:
```
running 2 tests
test bench_add_to_kdtree_with_1k_3d_points       ... bench:         139 ns/iter (+/- 61)
test bench_nearest_from_kdtree_with_1k_3d_points ... bench:       6,677 ns/iter (+/- 6,613)
```

Maby evan "0.3.1"?