@Eh2406 2016-07-25
HI,

The readme could be updated to mention the 3.x change, and the readme benchmarks may nead to be updated.

Thank you for making this crate and letting some random guy on the internet help with it. I've had fun working on it. I hope I have not been annoying. :-)